refactor: ImportSettings 관련 경로를 DEPRECATED 디렉터리로 이동
